+++ lib/main.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:resonance_network_wallet/core/extensions/color_extensions.dart';
+import 'package:resonance_network_wallet/features/main/screens/wallet_initializer.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 import 'package:qr_flutter/qr_flutter.dart';
 import 'dart:async';
@@ -140,160 +141,6 @@ class ResonanceWalletApp extends StatelessWidget {
       ),
       themeMode: ThemeMode.dark,
       home: const WalletInitializer(),
-    );
-  }
-}
-
-class WalletInitializer extends StatefulWidget {
-  const WalletInitializer({super.key});
-
-  @override
-  WalletInitializerState createState() => WalletInitializerState();
-}
-
-class WalletInitializerState extends State<WalletInitializer> {
-  bool _loading = true;
-  bool _walletExists = false;
-
-  @override
-  void initState() {
-    super.initState();
-    _checkWalletExists();
-  }
-
-  Future<void> _checkWalletExists() async {
-    final prefs = await SharedPreferences.getInstance();
-    final hasWallet = prefs.getBool('has_wallet') ?? false;
-
-    setState(() {
-      _walletExists = hasWallet;
-      _loading = false;
-    });
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    if (_loading) {
-      return const Scaffold(
-        body: Center(
-          child: CircularProgressIndicator(),
-        ),
-      );
-    }
-
-    if (_walletExists) {
-      return const WalletMain();
-    } else {
-      return const WelcomeScreen();
-    }
-  }
-}
-
-class WelcomeScreen extends StatelessWidget {
-  const WelcomeScreen({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      body: Container(
-        padding: const EdgeInsets.symmetric(horizontal: 30),
-        decoration: BoxDecoration(
-          gradient: LinearGradient(
-            begin: Alignment.topCenter,
-            end: Alignment.bottomCenter,
-            colors: [const Color(0xFF1A1A1A), const Color(0xFF6B46C1).useOpacity(0.8)],
-          ),
-        ),
-        child: Column(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: [
-            Container(
-              padding: const EdgeInsets.all(20),
-              decoration: BoxDecoration(
-                shape: BoxShape.circle,
-                color: const Color(0xFF2D2D2D),
-                boxShadow: [
-                  BoxShadow(
-                    color: const Color(0xFF6B46C1).useOpacity(0.3),
-                    blurRadius: 20,
-                    spreadRadius: 5,
-                  ),
-                ],
-              ),
-              child: const Icon(
-                Icons.account_balance_wallet,
-                size: 80,
-                color: Color(0xFF9F7AEA),
-              ),
-            ),
-            const SizedBox(height: 30),
-            const Text(
-              'Resonance Network',
-              style: TextStyle(
-                fontSize: 28,
-                fontWeight: FontWeight.bold,
-                color: Colors.white,
-              ),
-            ),
-            const SizedBox(height: 10),
-            const Text(
-              'Welcome to your REZ wallet',
-              style: TextStyle(
-                fontSize: 18,
-                color: Colors.white70,
-              ),
-            ),
-            const SizedBox(height: 60),
-            ElevatedButton(
-              onPressed: () {
-                Navigator.push(
-                  context,
-                  MaterialPageRoute(builder: (context) => const CreateWalletScreen()),
-                );
-              },
-              style: ElevatedButton.styleFrom(
-                backgroundColor: const Color(0xFF6B46C1),
-                minimumSize: const Size(double.infinity, 50),
-                shape: RoundedRectangleBorder(
-                  borderRadius: BorderRadius.circular(12),
-                ),
-              ),
-              child: const Text(
-                'Create New Wallet',
-                style: TextStyle(
-                  fontSize: 16,
-                  fontWeight: FontWeight.bold,
-                  color: Colors.white,
-                ),
-              ),
-            ),
-            const SizedBox(height: 20),
-            OutlinedButton(
-              onPressed: () {
-                Navigator.push(
-                  context,
-                  MaterialPageRoute(builder: (context) => const ImportWalletScreen()),
-                );
-              },
-              style: OutlinedButton.styleFrom(
-                side: const BorderSide(color: Color(0xFF9F7AEA)),
-                minimumSize: const Size(double.infinity, 50),
-                shape: RoundedRectangleBorder(
-                  borderRadius: BorderRadius.circular(12),
-                ),
-              ),
-              child: const Text(
-                'Import Existing Wallet',
-                style: TextStyle(
-                  fontSize: 16,
-                  fontWeight: FontWeight.bold,
-                  color: Color(0xFF9F7AEA),
-                ),
-              ),
-            ),
-          ],
-        ),
-      ),
     );
   }
 }
